test: consolidate subgraph test utils into SubgraphHelper
Move all standalone functions from helpers/subgraphTestUtils.ts into
SubgraphHelper as instance methods (serializeAndReload,
convertDefaultKSamplerToSubgraph, packAllInteriorNodes) and static
methods (getTextSlotPosition, expectWidgetBelowHeader,
collectConsoleWarnings). Update 9 spec files and delete the old module.

Amp-Thread-ID: https://ampcode.com/threads/T-019d3158-4e9c-76c8-84ae-dffccccbd0ca
