+++ browser_tests/tests/subgraphNestedConfigureOrder.spec.ts
@@ -1,7 +1,7 @@
 import { expect } from '@playwright/test'
 
 import { comfyPageFixture as test } from '../fixtures/ComfyPage'
-import { collectConsoleWarnings } from '../helpers/subgraphTestUtils'
+import { SubgraphHelper } from '../fixtures/helpers/SubgraphHelper'
 
 test.describe('Nested subgraph configure order', { tag: ['@subgraph'] }, () => {
   const WORKFLOW = 'subgraphs/subgraph-nested-duplicate-ids'
@@ -9,7 +9,7 @@ test.describe('Nested subgraph configure order', { tag: ['@subgraph'] }, () => {
   test('Loads without "No link found" or "Failed to resolve legacy -1" console warnings', async ({
     comfyPage
   }) => {
-    const { warnings } = collectConsoleWarnings(comfyPage.page)
+    const { warnings } = SubgraphHelper.collectConsoleWarnings(comfyPage.page)
 
     await comfyPage.workflow.loadWorkflow(WORKFLOW)
